refactor(plugins): rename plugin helpers file
