+++ src/lib/plugin-helpers.ts
